+++ CHANGELOG.md
@@ -3,8 +3,11 @@
 ## Development
 
 - 0.1.0 Functional Prototype
-> Working 'wireframe' prototype without functioning database. Moved groceries.json into `./routes/index.js` for testing purposes. No styling.
+>Working 'wireframe' prototype without functioning database. Moved groceries.json into `./routes/index.js` for testing purposes. No styling.
+
 - 0.2.0 Functional Back-End
 >Completed database CRUD methods and Groceries Schema. Adds the ability to render the current grocery list to handlebars index. This will be used to enable looping the active list.
 
+- 0.2.1 Modified Schema and db reset functionality
+>Modified Schema to make the combined index of `"name", "type", "brand"` fields unique. Added seeds.json and created controller methods allowing for easier reset of development db (testing purposes).